WIP: undo renaming action option in filter-action-row.tsx and fix dependency cycle and
- create-filter-modal was failing because of renaming so I fixed that and introduced a new type
